refactor: split groups.py into modular package structure (#180)
Refactored the 2073-line groups.py monolith into a proper package with
focused submodules:

- constants.py: Protocol constants and key sizes
- exceptions.py: MLSError and subclasses
- types.py: GroupRole, ProposalType, MemberStatus, GroupStatus enums
- key_package.py: KeyPackage class
- members.py: GroupMember, EpochSecrets classes
- messages.py: WelcomeMessage, CommitMessage, RemovalAuditEntry
- state.py: GroupState class
- operations.py: Group operations (create, add, remove, etc.)
- federation.py: FederationGroup and helpers
- storage.py: In-memory storage for MVP
- __init__.py: Re-exports public API for backward compatibility

No functional changes - all existing imports continue to work.
All 1158 federation tests pass.

Closes #180
